Changes
-Added partial global statistics, so now it needs to sync with all the threads everytime it reaches M combinations

-Extracted code which can be re-used (statisticsSummary) and changed the final synchronization method to adapt this new change

 -Fixed some statistics calculations that were wrong, and fixed critical bugs related to statistics that made the program crash depending on the configuration

 -TODO: fix averages and change header of statistics prints

+++ manfut.h
@@ -61,8 +61,8 @@ struct Tstatistics{
     int numComb;
     int numInvComb;
     int numValidComb;
-    int avgCostValidComb;
-    int avgScoreValidComb;
+    float avgCostValidComb;
+    float avgScoreValidComb;
     TJugadorsEquip bestCombination;
     int bestScore;
     TJugadorsEquip worseCombination;
